fix(api): prefer 'latest' blockTag when a timestamp has not been specified
The error that appears is a race condition at the Infura node level. When the caller omits a timestamp, the latest block (from node A) gets resolved. The function then uses this block to call various contract functions (from nodes B, C, and D). If node A reports a higher block before nodes B, C, or D observe the block, then an out-of-bounds error happens. This commit aims to resolve that by having the nodes use whatever block is their latest one in the common case (when a timestamp is omited).