feat(api): expose three new fields from response
The total fees from the relayFeeCalculator are also dependent on the current price of the token. In most cases, this is a non-issue because ETH is the base price. However, in the singular case of MATIC, the total amount must also be multiplied by the price of MATIC/ETH. To prevent needing this value to be calculated on the FE, this value can be exposed in the JSON payload of /suggested-fees