feat(api): add optional query param to skip Too Low error (#248)
* feat(api): add optional query param to skip Too Low error

This change creates a query parameter that skips the Amount Too Low error and returns its as part of the payload json.

* improve: remove additional variable requirement

* fix: refine scope of skipAmountLimit

* fix(api): test parameter specifically for `true`

This endpoint currently tests for the existence of the skipAmountLimit query. However, passing ?skipAmountLimit=false would still resolve to a non-undefined value. This fix tightens the scope of this query to only activating when the isAmountTooLow = true

* fix(api): always return isAmountTooLow in JSON response

This commit adds the `isAmountTooLow` parameter to all successful JSON responses

This should not cause any breaking changes. The JSON scope is expanded, but not reduced.