fix: port relayer fee logic to API (#256)
* feat: add start of api utility to call endpoints

* refactor: add fromChain to API calls

* feat: add start of api utility to call endpoints

* refactor: add fromChain to API calls

* feat: replace sdk call with api reference

* debug: add temp logging

* fix: update fixed point division

* feat(api): expose three new fields from response

The total fees from the relayFeeCalculator are also dependent on the current price of the token. In most cases, this is a non-issue because ETH is the base price. However, in the singular case of MATIC, the total amount must also be multiplied by the price of MATIC/ETH. To prevent needing this value to be calculated on the FE, this value can be exposed in the JSON payload of /suggested-fees

* feat: resolve new data fields from the API

* improve: add too low data from api

* fix(api): remove errant price cache

Before this commit, the cached CG price was being used as a overrided value to the getRelayerfeeDetails. This works for all the chains whose basis currency is ETH. However, this will fail for MATIC, because the price in the PolygonQueries is determined by a basis currency of MATIC.

* docs: add relevant comments
* fix(test): Add env to mock serverless API
* feat: add mock and prod api function designation