fix(limits): Fix minDeposit derivation on Polygon
The limits API endpoint assumes that the L1 (bridged) token is priced
in the native gas token. getCachedTokenPrice() defaults to returning
prices in ETH terms when baseCurrency is not specified.

When pricing the bridged token in ETH terms on Polygon, it incorrectly
devalues the bridged token and therefore produces an artificially high
minimum deposit amount.

Issue initially reported by the curious minds over at Rango Exchange.