feat: :sparkles: allow pool deep linking
This change enables deep linking on the pool page. A user can now specify the specific symbol that will be loaded by using the ?symbol={symbol} parameter. This parameter is case-insensitive and represents a token symbol such as ETH, WETH, UDSC, ACX, etc. This check is only performed at render time and is intended to be leveraged by the Rewards page to link the `Add` button to a specific pool.

Breaking Changes: N/A