feat: :sparkles: preserve position size when user stakes
This change allows the position size of the user's LP pool balance to remain the same even when they stake. Currently, when a user stakes with the AcceleratingDistributor contract their tokens are transferred to the contract - this reduces the number of tokens visible on the pool page. This change accounts for this difference.

Breaking Changes: N/A