improve: :sparkles: introduce multiple UI fixes (#464)
* fix: :zap: hide specific pool table if no elements are available

This change selectively hides certain pool tables if the loading has completed and there are no pool entries, the table is hidden

* fix: temporarily provide a default acx price

* fix: :bug: Add a decimal place to the multiplier staking field

* fix: :speech_balloon: update copy on splash screen

This change aligns the minimum transfer time on splash screen to the dynamic data

* feat: :sparkles: preserve position size when user stakes

This change allows the position size of the user's LP pool balance to remain the same even when they stake. Currently, when a user stakes with the AcceleratingDistributor contract their tokens are transferred to the contract - this reduces the number of tokens visible on the pool page. This change accounts for this difference.

Breaking Changes: N/A

* fix: make code context-aware of cypress

* improve: bundle acxprice loading into pool query

* feat: a lookup address to price goerli assets

* fix: enforce a string from coingecko api

* feat: accurately compute exchange price of tokens

* fix: preserve EOF

* fix: remove unneeded space