fix: :zap: replaces backup contract addresses with mainnet as default
This change sets the mainnet contract addresses for the AD, ACX token, Merkle Distributor, and the Claim & Stake as the default addresses if a valid config is not provided.

Breaking changes: this should not break anything. We have set this edge case for the event that the config file is invalid.