improve(limits): Dynamically lookup LP cushion env vars
Considered some of the feedback from James and Nick and decideded to try
something less computationally-intensive. This change now resolves the
L1 token symbol and dynamically looks up the LP cushion env var.

It does involve a search over the l1Tokens object; it would be nice to
avoid that, but I do think it's a worthwhile sacrifice for the reduction
in code size (in terms of LoC), and the marginal ease of adding padding
for new tokens.