fix: :bug: prevent crash on invalid route
This fix provides a check when a user enters an invalid route state. An invalid route state can be achieved when a user selects a token, chooses a destination/origin route, and then switches tokens that don't exist on the given origin/destination route. This change accounts for this by doing a safety check before the user changes the token inside the handler itself.