refactor: :fire: remove boba from FE
This change effectively removes the Boba Network/ Boba Token from the front end. Boba will no longer appear in any copy or in any routes that are enabled by Across. This is enforced at both the front end and in the contracts. No RPC calls to the Boba network will be made.