fix(ui): :ambulance: resolve autoconnect issue (#691)
This change corrects for the autoconnect functionality that was occasionally breaking the FE. This new change adds a new option to the `customOnboardConnect` function to enable the autoconnect functionality at all. This is a correction from the previous logic because in production currently the autoconnect tries to keep silentely failing causing some users to not be able to connect.