feat: :art: only reset bridge input on chain switch
This change disallows the input to be automatically reset when a chain is being switched. As such, when a user attempts to change chains, their input token size will not be changed. However, this may cause invalid inputs if the user switches to a chain where the input token balance is not high enough to bridge.

Breaking Changes: none