refactor: introduce state management for the quoteRequest in swap-ui (#1984)
This refactor extracts quote request state into a dedicated reducer pattern (useQuoteRequest), simplifying useSwapAndBridge to be a pure computation hook.

This is the first of many refactoring PRs, aimed at simplifying the implementation and making it easier to understand.

My goal here was not to get too deep into the different components, but to focus on the InputForm and SwapAndBridge components and the useSwapAndBridge hook