feat: add feature flag capabilities to across app (#1935)
We want to be able to toggle features on and off for specific users with Amplitude experiment. 

This PR introduces a mechanism for initialising feature flags across the frontend with the relevant user data. The feature flags are fetched on application load and when the wallet provider changes the connected wallet. The feature flags are also cached in localstorage, so that subsequent refreshes will be faster. 

This is the first small step towards feature flags. In the future, we might want to
* support feature flag payloads so that properties like text, colours, and images can be changed from amplitude without requiring a commit. 
* Improve speed by caching feature flags in vercel functions on the edge


https://github.com/user-attachments/assets/f60f5a66-28f8-457c-bc7b-86b316a340c3

feature flag configuration page: https://app.amplitude.com/experiment/risklabs/414806/config/680059/configure